pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "konflux-automation"
version = "0.1.0"
description = "Automation toolkit for Konflux operations"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "Konflux Automation" }
]
license = { text = "Apache-2.0" }
dependencies = [
  "kubernetes>=27.2.0",
  "typer>=0.9.0",
  "pydantic>=2.5.0",
  "PyYAML>=6.0",
  "rich>=13.0",
  "ruamel.yaml>=0.18"
]

[project.urls]
Homepage = "https://example.com/konflux-automation"

[project.scripts]
konflux-ops = "konflux_automation.cli:app"
